Fixed build error reported by ICC and GCC


git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@12209 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Bus/Pci/XhciDxe/Xhci.c
@@ -829,7 +829,7 @@ XhcControlTransfer (
   // endpoint is bidirectional. XhcCreateUrb expects this
   // combination of Ep addr and its direction.
   //
-  Endpoint = 0 | ((TransferDirection == EfiUsbDataIn) ? 0x80 : 0);
+  Endpoint = (UINT8) (0 | ((TransferDirection == EfiUsbDataIn) ? 0x80 : 0));
   Urb = XhcCreateUrb (
           Xhc,
           XhciDevAddr,
